requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# MCP Protocol
mcp>=0.1.0

# Configuration and utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Note: llmwrapper and schema-graph-builder should be installed with:
# pip install -e ../llmwrapper
# pip install -e ../schema-graph-builder 